+++ src/main/resources/sqlScripts/postgres/schema-postgres.sql
@@ -24,8 +24,8 @@ CREATE TABLE world_of_manga.role (
 DROP TABLE IF EXISTS world_of_manga.user_role CASCADE;
 
 CREATE TABLE world_of_manga.user_role(
-	"user_id" int,
-    "role_id" int,
+	"user_id" INT,
+    "role_id" INT,
     PRIMARY KEY("user_id", "role_id"),
 	CONSTRAINT "FK_USER_ROLE" FOREIGN KEY ("user_id") REFERENCES world_of_manga.user("user_id") ON DELETE NO ACTION ON UPDATE NO ACTION,
 	CONSTRAINT "FK_ROLE_USER" FOREIGN KEY ("role_id") REFERENCES world_of_manga.role("role_id") ON DELETE NO ACTION ON UPDATE NO ACTION
@@ -70,8 +70,8 @@ DROP TABLE IF EXISTS world_of_manga.manga_translation CASCADE;
 
 CREATE TABLE world_of_manga.manga_translation (
     "manga_translation_id" SERIAL PRIMARY KEY,
-    "manga_id" int,
-    "language_id" int,
+    "manga_id" INT,
+    "language_id" INT,
     "title" VARCHAR(50) NOT NULL,
     "description" VARCHAR(1000) NOT NULL,
     CONSTRAINT "FK_MANGA_TRANSLATION" FOREIGN KEY ("manga_id")
@@ -87,7 +87,7 @@ DROP TABLE IF EXISTS world_of_manga.manga_rating CASCADE;
 CREATE TABLE world_of_manga.manga_rating (
     "manga_id" SERIAL,
     "user_id" SERIAL,
-    "rating" DECIMAL(3 , 2 ),
+    "rating" INT,
     PRIMARY KEY ("manga_id" , "user_id"),
     CONSTRAINT "FK_RATING_MANGA_USER" FOREIGN KEY ("manga_id")
         REFERENCES world_of_manga.manga ("manga_id")
